+++ provider/metrics_test.go
@@ -1,4 +1,4 @@
-package azure
+package provider
 
 import (
 	"path"
